refactor: restructure a little and add tests
