+++ pyproject.toml
@@ -59,7 +59,7 @@ Discussions = "https://github.com/scikit-build/scikit-build-cli/discussions"
 Changelog = "https://github.com/scikit-build/scikit-build-cli/releases"
 
 [project.scripts]
-skbuild = "scikit_build_cli.__main__:run_cli"
+skbuild = "scikit_build_cli.__main__:skbuild"
 
 [project.entry-points."skbuild.commands"]
 build = "scikit_build_cli.commands.build:build"
